backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
sqlalchemy==2.0.36
websockets==14.1
pydantic==2.10.6
python-dotenv==1.0.1
